+++ glade_forms/example/lib/example.dart
@@ -12,9 +12,9 @@ class _Model extends GladeModel {
 
   @override
   void initialize() {
-    name = GladeInput.stringInput();
-    age = GladeInput.intInput(value: 0);
-    email = GladeInput.stringInput(validator: (validator) => (validator..isEmail()).build());
+    name = GladeInput.stringInput(inputKey: 'name');
+    age = GladeInput.intInput(value: 0, inputKey: 'age');
+    email = GladeInput.stringInput(validator: (validator) => (validator..isEmail()).build(), inputKey: 'email');
 
     super.initialize();
   }
@@ -25,7 +25,7 @@ class Example extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return GladeFormBuilder(
+    return GladeFormBuilder.create(
       create: (context) => _Model(),
       builder: (context, model, _) => Padding(
         padding: const EdgeInsets.all(32),