+++ glade_forms/lib/src/model/glade_model.dart
@@ -3,6 +3,9 @@ import 'package:glade_forms/src/core/core.dart';
 import 'package:meta/meta.dart';
 
 abstract class GladeModel extends ChangeNotifier {
+  List<GladeInput<Object?>> _lastUpdates = [];
+  bool _groupEdit = false;
+
   bool get isValid => inputs.every((input) => input.isValid);
 
   bool get isNotValid => !isValid;
@@ -14,6 +17,9 @@ abstract class GladeModel extends ChangeNotifier {
   bool get isDirty => !isPure;
 
   List<GladeInput<Object?>> get inputs;
+
+  List<String> get lastUpdatedInputKeys =>
+      _lastUpdates.map((e) => e.inputKey).where((element) => element != null).cast<String>().toList();
 
   /// Formats errors from `inputs`.
   String get formattedValidationErrors => inputs.map((e) {
@@ -56,10 +62,33 @@ abstract class GladeModel extends ChangeNotifier {
   void updateInput<INPUT extends GladeInput<T?>, T>(INPUT input, T value) {
     if (input.value == value) return;
 
+    _lastUpdates = [input];
+
     input.value = value;
     notifyListeners();
   }
 
   @internal
-  void notifyInputUpdated() => notifyListeners();
+  void notifyInputUpdated(GladeInput<Object?> input) {
+    if (_groupEdit) {
+      _lastUpdates.add(input);
+    } else {
+      _lastUpdates = [input];
+    }
+
+    if (!_groupEdit) {
+      notifyListeners();
+    }
+  }
+
+  /// Use it to update multiple inputs at once before these changes are popragated through notifyListeners().
+  void groupEdit(void Function() edit) {
+    _groupEdit = true;
+
+    edit();
+
+    _groupEdit = false;
+
+    notifyListeners();
+  }
 }